requirements.txt
Django==2.2
django-cors-headers==3.4.0
djangorestframework==3.11.0
requests==2.18.1
